Update README.md
A to-do list application is a productivity tool that assists users in methodically organizing their projects, to-do lists, and reminders. Tasks can be added, changed, rearranged, and deleted by users as needed. Adding due dates, creating reminders, classifying tasks into various lists or categories, and even sharing tasks with others are just a few of the functions that the application usually offers. To keep organized and increase productivity, to-do list software is extensively used for team collaboration, project planning, and individual task management.